pyproject.toml
[project]
name = "epo-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "tenacity>=9.1.4",
]
